requirements.txt
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bertopic==0.16.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-cli==0.18.0
databricks-sdk==0.77.0
docker==6.1.3
entrypoints==0.4
fastapi==0.109.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.47.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httpx==0.26.0
huey==2.6.0
huggingface-hub==0.20.3
idna==3.11
importlib_metadata==7.2.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.45.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mpmath==1.3.0
narwhals==2.12.0
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==1.26.3
oauthlib==3.3.1
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==23.2
pandas==2.1.4
pillow==10.4.0
plotly==5.18.0
pluggy==1.6.0
protobuf==6.33.3
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.5
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
pytz==2023.4
pywin32==311
PyYAML==6.0.3
querystring-parser==1.2.4
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.3.2
scipy==1.16.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.35.1
streamlit==1.30.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
toml==0.10.2
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tornado==6.5.2
tqdm==4.67.1
transformers==4.40.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
urllib3==2.5.0
uvicorn==0.27.0
validators==0.35.0
waitress==2.1.2
watchdog==6.0.0
websocket-client==1.9.0
Werkzeug==3.1.3
zipp==3.23.0
